Update new project

+++ edu-exam/02-backend/src/main/java/com/wewe/eduexam/service/StudentExamResultService.java
@@ -5,6 +5,6 @@ import com.wewe.eduexam.model.StudentExamResult;
 import java.util.Optional;
 
 public interface StudentExamResultService {
-    StudentExamResult saveResult(Long studentId, Long examId, Integer totalScore);
+    StudentExamResult saveResult(Long studentId, Long examId, Double totalScore);
     Optional<StudentExamResult> findById(Long resultId);
 }
